refactor: extract DrillPathAccordionItem and FieldMappingEditor from DrillConfigPanel
Split the 1100-line DrillConfigPanel into three focused files to reduce
nesting depth and improve readability. No behavioral changes.
